Resolve build errors on GenomicsReport
Need to change how we define sliceName (Profile Name column) and path (Element column) to resolve the new build errors.

NOTE - It is likely this is not the best (or even right) way to do this, but it resolves some of our new build failures.  Will need to review closely.